azure_function/order_extraction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
psycopg2-binary
pandas
openpyxl
python-docx
openai
azure-storage-blob
azure-identity
